+++ CHANGELOG.rst
@@ -2,9 +2,11 @@
 Changelog for package BebopS
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-0.2.1 (2019-XX-XX)
+0.2.1 (2019-02-12)
 ------------------
 * The interface between the position_controller and the Parrot Sphinx simulator has been made
+* Added the hovering example file to publish the waypoint when Sphinx is in the loop
+* Added the waypoint example file to publish waypoints when Sphinx is in the loop
 * Contributors: Giuseppe Silano, Peter Griggs
 
 0.2.0 (2019-01-31)